+++ src/components/Box/RvVStack.tsx
@@ -1,6 +1,6 @@
 import type { FC } from "react";
 import { StyledVStack } from "./style";
-import type { RvVStackFCType, StyledVStackType } from "../../types";
+import type { RvStackFCType, StyledVStackType } from "../../types";
 
 export const rvVStackDefaultProps: StyledVStackType = {
 	width: "100%",
@@ -9,7 +9,7 @@ export const rvVStackDefaultProps: StyledVStackType = {
 	marginBottom: 5,
 };
 
-const RvVStack: FC<RvVStackFCType> = ({ children, ...props }) => {
+const RvVStack: FC<RvStackFCType> = ({ children, ...props }) => {
 	const width = { ...props }.width ?? { ...rvVStackDefaultProps }.width;
 	const borderRadius = { ...props }.borderRadius ?? { ...rvVStackDefaultProps }.borderRadius;
 	const marginBottom = { ...props }.space ?? { ...rvVStackDefaultProps }.marginBottom;