+++ src/components/Box/index.ts
@@ -1,6 +1,7 @@
 import RvBox from "./RvBox";
 import RvFlex from "./RvFlex";
 import RvVStack from "./RvVStack";
+import RvHStack from "./RvHStack";
 
-export { RvBox, RvFlex, RvVStack };
+export { RvBox, RvFlex, RvVStack, RvHStack };
 export * from "./style";