+++ src/components/index.ts
@@ -1,3 +1,5 @@
 export * from "./Box";
 export * from "./Button";
+export * from "./Card";
+export * from "./Span";
 export * from "./config";