+++ src/types/components/index.ts
@@ -1,4 +1,4 @@
 export * from "./Div";
 export * from "./Button";
+export * from "./Card";
 export * from "./Span";
-export * from "../CSSProperties";